microservices/search/search/requirements.txt
fastapi
uvicorn
redis
python-dotenv
pytest
pymongo
python-Levenshtein
APScheduler
requests
